Move around scripting utils (#7523)
* Move around scripting utils

* Don't set a file path

* Version bump
